change Architecture to layers
