+++ app/src/main/java/com/example/dispatcher/common/base/BaseFragment.kt
@@ -4,7 +4,7 @@ import com.example.dispatcher.common.utils.JsonDataManager
 import android.os.Bundle
 import android.view.View
 import androidx.fragment.app.Fragment
-import com.example.dispatcher.features.homepage.presentation.model.Article
+import com.example.dispatcher.presentation.homepage.model.Article
 
 abstract class BaseFragment : Fragment() {
