+++ app/src/main/java/com/example/dispatcher/presentation/favorites/view/FavoritesFragment.kt
@@ -1,41 +1,43 @@
-package com.example.dispatcher.features.profilepage.presentation.view
+package com.example.dispatcher.presentation.favorites.view
 
 import android.os.Bundle
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import android.widget.TextView
-import com.example.dispatcher.databinding.FragmentProfileBinding
+import com.example.dispatcher.databinding.FragmentFavoritesBinding
 import com.example.dispatcher.common.base.BaseFragment
 import com.example.dispatcher.common.utils.displayToast
 
-class ProfileFragment : BaseFragment() {
+class FavoritesFragment : BaseFragment() {
 
-    private var _binding: FragmentProfileBinding? = null
+    private var _binding: FragmentFavoritesBinding? = null
     private val binding get() = _binding!!
 
     override fun onCreateView(
         inflater: LayoutInflater, container: ViewGroup?,
         savedInstanceState: Bundle?
     ): View? {
-        _binding = FragmentProfileBinding.inflate(inflater, container, false)
+        // Inflate the layout using view binding
+        _binding = FragmentFavoritesBinding.inflate(inflater, container, false)
         return binding.root
     }
 
     override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
         super.onViewCreated(view, savedInstanceState)
 
-        // Dynamically create and add a TextView for each author's name
-        articleList.mapNotNull { it.author }.forEach { author ->
+        // Dynamically create a TextView for each title and add it to the layout
+        articleList.mapNotNull { it.title }.forEach { title ->
             val textView = TextView(requireContext()).apply {
-                text = author
+                text = title
                 textSize = 18f
                 setTextColor(resources.getColor(android.R.color.black, null))
             }
+            // Add the TextView to the root LinearLayout
             binding.root.addView(textView)
         }
 
-        displayToast("Profile Fragment is active!")
+        displayToast("Favorites Fragment is active!")
     }
 
     override fun onDestroyView() {
