+++ app/src/main/java/com/example/dispatcher/presentation/homepage/view/HomeFragment.kt
@@ -1,43 +1,44 @@
-package com.example.dispatcher.features.favorites.presentation.view
+package com.example.dispatcher.presentation.homepage.view
 
 import android.os.Bundle
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import android.widget.TextView
-import com.example.dispatcher.databinding.FragmentFavoritesBinding
+import com.example.dispatcher.databinding.FragmentHomeBinding
 import com.example.dispatcher.common.base.BaseFragment
 import com.example.dispatcher.common.utils.displayToast
 
-class FavoritesFragment : BaseFragment() {
+class HomeFragment : BaseFragment() {
 
-    private var _binding: FragmentFavoritesBinding? = null
+    private var _binding: FragmentHomeBinding? = null
     private val binding get() = _binding!!
 
     override fun onCreateView(
         inflater: LayoutInflater, container: ViewGroup?,
         savedInstanceState: Bundle?
     ): View? {
-        // Inflate the layout using view binding
-        _binding = FragmentFavoritesBinding.inflate(inflater, container, false)
+        _binding = FragmentHomeBinding.inflate(inflater, container, false)
         return binding.root
     }
 
     override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
         super.onViewCreated(view, savedInstanceState)
 
-        // Dynamically create a TextView for each title and add it to the layout
-        articleList.mapNotNull { it.title }.forEach { title ->
+        // Dynamically create and add a TextView for the first two words of each article's body
+        articleList.mapNotNull { article ->
+            article.body?.split(" ")?.take(2)?.joinToString(" ")
+        }.forEach { firstTwoWords ->
             val textView = TextView(requireContext()).apply {
-                text = title
+                text = firstTwoWords
                 textSize = 18f
                 setTextColor(resources.getColor(android.R.color.black, null))
             }
-            // Add the TextView to the root LinearLayout
+
             binding.root.addView(textView)
         }
 
-        displayToast("Favorites Fragment is active!")
+        displayToast("Home Fragment is active!")
     }
 
     override fun onDestroyView() {
