+++ app/src/main/java/com/example/dispatcher/presentation/profile/view/ProfileFragment.kt
@@ -1,44 +1,41 @@
-package com.example.dispatcher.features.homepage.presentation.view
+package com.example.dispatcher.presentation.profile.view
 
 import android.os.Bundle
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import android.widget.TextView
-import com.example.dispatcher.databinding.FragmentHomeBinding
+import com.example.dispatcher.databinding.FragmentProfileBinding
 import com.example.dispatcher.common.base.BaseFragment
 import com.example.dispatcher.common.utils.displayToast
 
-class HomeFragment : BaseFragment() {
+class ProfileFragment : BaseFragment() {
 
-    private var _binding: FragmentHomeBinding? = null
+    private var _binding: FragmentProfileBinding? = null
     private val binding get() = _binding!!
 
     override fun onCreateView(
         inflater: LayoutInflater, container: ViewGroup?,
         savedInstanceState: Bundle?
     ): View? {
-        _binding = FragmentHomeBinding.inflate(inflater, container, false)
+        _binding = FragmentProfileBinding.inflate(inflater, container, false)
         return binding.root
     }
 
     override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
         super.onViewCreated(view, savedInstanceState)
 
-        // Dynamically create and add a TextView for the first two words of each article's body
-        articleList.mapNotNull { article ->
-            article.body?.split(" ")?.take(2)?.joinToString(" ")
-        }.forEach { firstTwoWords ->
+        // Dynamically create and add a TextView for each author's name
+        articleList.mapNotNull { it.author }.forEach { author ->
             val textView = TextView(requireContext()).apply {
-                text = firstTwoWords
+                text = author
                 textSize = 18f
                 setTextColor(resources.getColor(android.R.color.black, null))
             }
-
             binding.root.addView(textView)
         }
 
-        displayToast("Home Fragment is active!")
+        displayToast("Profile Fragment is active!")
     }
 
     override fun onDestroyView() {
