+++ app/src/main/res/layout/activity_main.xml
@@ -91,6 +91,6 @@
         android:layout_below="@id/headerFragmentContainer"
         android:layout_above="@id/tabsLayout"
         android:layout_centerInParent="true"
-        android:name="com.example.dispatcher.features.homepage.presentation.view.HomeFragment" />
+        android:name="com.example.dispatcher.presentation.homepage.view.HomeFragment" />
 
 </RelativeLayout>
